+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Fixed
 - article category on homepage
+- Item::related scope to match catalog search query
 
 ### Added
 - default Scout config to enable Scout search functionality